+++ software/NRG_itho_wifi/main/websocket.cpp
@@ -302,15 +302,15 @@ static void wsEvent(struct mg_connection *c, int ev, void *ev_data, void *fn_dat
               uint8_t index = root["ithosetupdate"].as<uint8_t>();
               if (ithoSettingsArray[index].type == ithoSettings::is_float && ithoSettingsArray[index].length == 1)
               {
-                updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int8_t>(root["value"].as<double>() * ithoSettingsArray[index].divider), true);
+                updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int8_t>(root["value"].as<float>() * ithoSettingsArray[index].divider), true);
               }
               else if (ithoSettingsArray[index].type == ithoSettings::is_float && ithoSettingsArray[index].length == 2)
               {
-                updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int16_t>(root["value"].as<double>() * ithoSettingsArray[index].divider), true);
+                updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int16_t>(root["value"].as<float>() * ithoSettingsArray[index].divider), true);
               }
               else if (ithoSettingsArray[index].type == ithoSettings::is_float && ithoSettingsArray[index].length == 4)
               {
-                updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int32_t>(root["value"].as<double>() * ithoSettingsArray[index].divider), true);
+                updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int32_t>(root["value"].as<float>() * ithoSettingsArray[index].divider), true);
               }
               else
               {
@@ -415,15 +415,15 @@ static void wsEvent(struct mg_connection *c, int ev, void *ev_data, void *fn_dat
           uint8_t index = root["ithosetupdate"].as<uint8_t>();
           if (ithoSettingsArray[index].type == ithoSettings::is_float && ithoSettingsArray[index].length == 1)
           {
-            updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int8_t>(root["value"].as<double>() * ithoSettingsArray[index].divider), false);
+            updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int8_t>(root["value"].as<float>() * ithoSettingsArray[index].divider), false);
           }
           else if (ithoSettingsArray[index].type == ithoSettings::is_float && ithoSettingsArray[index].length == 2)
           {
-            updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int16_t>(root["value"].as<double>() * ithoSettingsArray[index].divider), false);
+            updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int16_t>(root["value"].as<float>() * ithoSettingsArray[index].divider), false);
           }
           else if (ithoSettingsArray[index].type == ithoSettings::is_float && ithoSettingsArray[index].length == 4)
           {
-            updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int32_t>(root["value"].as<double>() * ithoSettingsArray[index].divider), false);
+            updateSetting(root["ithosetupdate"].as<uint8_t>(), static_cast<int32_t>(root["value"].as<float>() * ithoSettingsArray[index].divider), false);
           }
           else
           {
